MasterとPlayerのファイルを統合
